+++ src/kubedash/TODO.md
@@ -1,0 +1,8 @@
+# TODO
+
+* gunicorn
+  * termibal, and log
+* pytest
+* opentelemetry instrumentors
+  * tracing
+  * oben jager from UI
