+++ src/kubedash/lib_plugins/external_loadbalancer.py
@@ -3,20 +3,24 @@
 from flask import Blueprint, render_template, request, session, redirect, url_for
 from flask_login import login_required
 
-from functions.helper_functions import get_logger, ErrorHandler
-from functions.sso import get_user_token
-from functions.k8s import k8sClientConfigGet, k8sNamespaceListGet
+from lib_functions.helper_functions import get_logger, ErrorHandler
+from lib_functions.sso import get_user_token
+from lib_functions.k8s import k8sClientConfigGet, k8sNamespaceListGet
 
 from kubernetes.client.rest import ApiException
 import kubernetes.client as k8s_client
 import ast
 
 ##############################################################
-## Helper Functions
+## variables
 ##############################################################
 
 exlb_routes = Blueprint("external_loadbalancer", __name__)
-logger = get_logger(__name__)
+logger = get_logger(__name__.split(".")[1])
+
+##############################################################
+## Helper Functions
+##############################################################
 
 def ipaddresspoolTest(namespace):
     k8s_object_error, k8s_object_list = ipaddresspoolsGet(namespace)
