+++ src/kubedash/lib_routes/main.py
@@ -1,0 +1,170 @@
+import requests
+from flask import Blueprint, jsonify, render_template, session, flash, redirect, url_for, request, Response
+from flask_login import login_user, login_required, logout_user
+from werkzeug.security import check_password_hash
+from itsdangerous import base64_decode
+from datetime import datetime
+
+from lib_functions.helper_functions import get_logger
+from lib_functions.sso import SSOSererGet, get_auth_server_info
+from lib_functions.user import User, UsersRoles, Role, KubectlConfig, SSOTokenGet
+from lib_functions.k8s import k8sServerConfigGet
+
+##############################################################
+## Helpers
+##############################################################
+
+main = Blueprint("main", __name__)
+logger = get_logger(__name__.split(".")[1])
+
+##############################################################
+## Login
+##############################################################
+
+@main.route('/')
+def login():
+        is_sso_enabled = False
+        is_ldap_enabled = False
+        authorization_url = None
+
+        if request.environ.get('HTTP_X_FORWARDED_FOR') is None:
+            remote_addr = request.remote_addr
+        else:
+            remote_addr = request.environ['HTTP_X_FORWARDED_FOR']
+
+        ssoServer = SSOSererGet()
+        if ssoServer is not None:
+            auth_server_info, oauth = get_auth_server_info()
+            if auth_server_info is not None:
+                print(auth_server_info)
+                auth_url = auth_server_info["authorization_endpoint"]
+                authorization_url, state = oauth.authorization_url(
+                    auth_url,
+                    access_type="offline",  # not sure if it is actually always needed,
+                                            # may be a cargo-cult from Google-based example
+                )
+                session['oauth_state'] = state
+                is_sso_enabled = True
+            else:
+                is_sso_enabled = False
+                flash('Cannot connect to identity provider!', "error")
+
+        if "username" in session:
+            username = session["username"]
+            k8sConfig = k8sServerConfigGet()
+            if k8sConfig is None:
+                logger.error("Kubectl Integration is not configured.")
+            else:
+                logger.info("Kubectl Integration is configured.")
+                k8s_server_ca = str(base64_decode(k8sConfig.k8s_server_ca), 'UTF-8')
+                try:
+                    i = requests.get('http://%s:8080/info' % remote_addr)
+                    info = i.json()
+                    if info["message"] == "kdlogin":
+                        user = User.query.filter_by(username=username, user_type = "OpenID").first()
+                        user2 = KubectlConfig.query.filter_by(name=session['username']).first()
+                        if is_sso_enabled and user:
+                            token = eval(SSOTokenGet(username))
+                            x = requests.post('http://%s:8080/' % remote_addr, json={
+                                    "username": username,
+                                    "context": k8sConfig.k8s_context,
+                                    "server": k8sConfig.k8s_server_url,
+                                    "certificate-authority-data": k8s_server_ca,
+                                    "client-id": ssoServer.client_id,
+                                    "id-token": token.get("id_token"),
+                                    "refresh-token": token.get("refresh_token"),
+                                    "idp-issuer-url": ssoServer.oauth_server_uri,
+                                    "client_secret": ssoServer.client_secret,
+                                }
+                            )
+                        elif user2:
+                            user_private_key = str(base64_decode(user2.private_key), 'UTF-8')
+                            user_certificate = str(base64_decode(user2.user_certificate), 'UTF-8')
+                            x = requests.post('http://%s:8080/' % remote_addr, json={
+                                    "username": username,
+                                    "context": k8sConfig.k8s_context,
+                                    "server": k8sConfig.k8s_server_url,
+                                    "certificate-authority-data": k8s_server_ca,
+                                    "user-private-key": user_private_key,
+                                    "user-certificate": user_certificate,
+                                }
+                            )
+                        logger.info("Config sent to client")
+                        logger.info("Answer from clinet: %s" % x.text)
+                except:
+                    pass
+            return redirect(url_for('main.cluster_metrics'))
+        else:
+            return render_template(
+                'login.html.j2',
+                sso_enabled = is_sso_enabled,
+                ldap_enabled = is_ldap_enabled,
+                auth_url = authorization_url
+            )
+        
+@main.route('/', methods=['POST'])
+def login_post():
+    username = request.form.get('username')
+    password = request.form.get('password')
+    remember = True if request.form.get('remember') else False
+
+    if request.environ.get('HTTP_X_FORWARDED_FOR') is None:
+        remote_addr = request.remote_addr
+    else:
+        remote_addr = request.environ['HTTP_X_FORWARDED_FOR']
+
+    user = User.query.filter(User.username == username, User.user_type != "OpenID").first()
+    user2 = KubectlConfig.query.filter_by(name=username).first()
+
+    # check if user actually exists
+    # take the user supplied password, hash it, and compare it to the hashed password in database
+    if not user or not check_password_hash(user.password_hash, password):
+        flash('Please check your login details and try again.', "warning")
+        return redirect(url_for('main.login')) # if user doesn't exist or password is wrong, reload the page
+    else:
+        user_role = UsersRoles.query.filter_by(user_id=user.id).first()
+        role = Role.query.filter_by(id=user_role.role_id).first()
+        login_user(user, remember=remember)
+        session['username'] = username
+        session['user_role'] = role.name
+        session['user_type'] = user.user_type
+        session['ns_select'] = "default"
+
+        k8sConfig = k8sServerConfigGet()
+        if k8sConfig is None:
+            logger.error ("Kubectl Integration is not configured.")
+        else:
+            k8s_server_ca = str(base64_decode(k8sConfig.k8s_server_ca), 'UTF-8')
+            try:
+                i = requests.get('http://%s:8080/info' % remote_addr)
+                info = i.json()
+                if info["message"] == "kdlogin" and user2:
+                    user_private_key = str(base64_decode(user2.private_key), 'UTF-8')
+                    user_certificate = str(base64_decode(user2.user_certificate), 'UTF-8')
+                    x = requests.post('http://%s:8080/' % remote_addr, json={
+                            "username": username,
+                            "context": k8sConfig.k8s_context,
+                            "server": k8sConfig.k8s_server_url,
+                            "certificate-authority-data": k8s_server_ca,
+                            "user-private-key": user_private_key,
+                            "user-certificate": user_certificate,
+                        }
+                    )
+                    logger.info("Config sent to client")
+                    logger.info("Answer from clinet: %s" % x.text)
+            except:
+                pass
+
+        return redirect(url_for('main.cluster_metrics'))
+
+
+@main.route('/logout')
+@login_required
+def logout():
+    logout_user()
+    if "username" in session:
+        session.pop('username', None)
+    if "oauth_token" in session:
+        session.pop('oauth_token')
+    session.clear()
+    return redirect(url_for('main.login'))
