+++ src/kubedash/requirements.txt
@@ -17,8 +17,9 @@ flask_sqlalchemy==3.0.5
 flask_login==0.6.3
 flask_session==0.8.0
 requests_oauthlib==2.0.0
-Flask_WTF==1.2.1
 # web security
+Flask_WTF==1.2.1
+flask_cors==5.0.0
 flask_talisman==1.1.0
 # kubernetes
 kubernetes==26.1.0
@@ -39,7 +40,6 @@ opentelemetry-semantic-conventions==0.44b0
 opentelemetry-instrumentation-sqlalchemy==0.44b0
 opentelemetry-instrumentation-flask==0.44b0
 # monitoring
-#prometheus_client==0.20.0
 prometheus-flask-exporter==0.23.1
 
 # security
@@ -49,7 +49,7 @@ setuptools==70.0.0
 pyOpenSSL==24.2.1
 cryptography==43.0.1
 ipython==8.27.0 # not directly required, pinned by Snyk to avoid a vulnerability
-gevent>=23.9.0 # not directly required, pinned by Snyk to avoid a vulnerability
-urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
-requests>=2.32.2 # not directly required, pinned by Snyk to avoid a vulnerability
-zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
+gevent==23.9.0 # not directly required, pinned by Snyk to avoid a vulnerability
+urllib3==2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
+requests==2.32.2 # not directly required, pinned by Snyk to avoid a vulnerability
+zipp==3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
