+++ src/kubedash/templates/400.html.j2
@@ -57,14 +57,14 @@
     </a>
   
     <!-- Bootstrap core JavaScript-->
-    <script src="static/vendor/jquery/jquery.min.js"></script>
-    <script src="static/vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
-  
+    <script src="{{ url_for('static',filename='/vendor/jquery/jquery.min.js') }}"></script>
+    <script src="{{ url_for('static',filename='/vendor/bootstrap/js/bootstrap.bundle.min.js') }}"></script> 
+
     <!-- Core plugin JavaScript-->
-    <script src="static/vendor/jquery-easing/jquery.easing.min.js"></script>
-  
+    <script src="{{ url_for('static',filename='/vendor/jquery-easing/jquery.easing.min.js') }}"></script>
+
     <!-- Custom scripts for all pages-->
-    <script src="static/js/sb-admin-2.min.js"></script>
+    <script src="{{ url_for('static',filename='/js/sb-admin-2.min.js') }}"></script> 
   
     </body>
 </html>
