+++ src/kubedash/tests/conftest.py
@@ -1,10 +1,16 @@
 import pytest, logging
+import flask_migrate
+
 from flask.testing import FlaskClient
 from kubedash import create_app
 
 @pytest.fixture(scope='module')
 def app():
     app = create_app("testing")
+
+    with app.app_context():
+        flask_migrate.upgrade()
+        
     yield app
 
 @pytest.fixture(scope='module')
